app/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
PyMySQL==1.0.2
python-dotenv==1.0.0
cryptography==41.0.5
